mpegmap/HTMLSourcingInbandTracks-3.5.b-1.html

<!doctype html>
<html>
  <head>
    <title>TC 5-3 Backward Seek Closed Caption Cues</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS, Closed Caption Cues - backward seek, UA does not create cues.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The media is an MPEG-2, Single Program Transport Stream
         containing both 608 and 708 Closed Caption.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>

    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>
    <script>
      var test = async_test("Backward Seek Closed Caption Cues.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "ClosedCaption.ts"; 

      var firstPassCues = null;
      var secondPassCues = null;
      var kind = "captions";
      var track = null;

      vid.textTracks.addEventListener("addtrack", function(event) {
        if (event.track.kind === kind) {
          track = event.track;
          track.mode = "hidden";
          track.addEventListener("cuechange", function() {
            if (!firstPassCues) {
              firstPassCues = [];
              for (var i = 0; i < track.cues.length; i++) {
                firstPassCues[i] = track.cues[i];
              }
              vid.currentTime = 0;
            }
          });
        }
      });

      setTimeout(function() {
        if (!track) {
          vid.pause();
          test.step(function() {
            assert_unreached("Did not find Closed Caption Track (kind) " +
                             kind + " in DOM.");
          }, "Did not find Closed Caption Track (kind) " + kind + " in DOM.");
          test.done();
        }
      }, 2000);

      vid.addEventListener("seeked", function() {
        track.addEventListener("cuechange", function() {
          if (!secondPassCues) {
            secondPassCues = [];
            for (var i = 0; i < track.cues.length; i++) {
              secondPassCues[i] = track.cues[i];
            }

            test.step(function() {
              assert_not_equals(firstPassCues.length, 0,
                                "First play, Closed Caption, TextTrackCues.length was 0.");
              assert_not_equals(secondPassCues.length, 0,
                                "Second play, Closed Caption, TextTrackCues.length was 0.");
              for (var i = 0; i < firstPassCues.length && i < secondPassCues.length; i++) {
                  assert_equals(secondPassCues[i].text, firstPassCues[i].text,
                                    'Cue[' + i + ']' + ' was different the second time.');
              }
            });
            test.done();
            vid.pause();
          }
        });
      });
    </script>
  </body>
</html>
